refatorado modulo database para dentro do infra
