+++ food-infra/src/main/java/br/com/alfac/food/infra/item/mapper/ItemEntityMapper.java
@@ -1,0 +1,14 @@
+package br.com.alfac.food.infra.item.mapper;
+
+import br.com.alfac.food.core.domain.item.Item;
+import br.com.alfac.food.infra.item.persistence.ItemEntity;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface ItemEntityMapper {
+
+    ItemEntity toEntity(Item item);
+
+    Item toDomain(ItemEntity Item);
+
+}
