+++ food-infra/src/main/java/br/com/alfac/food/infra/item/persistence/ItemEntityRepository.java
@@ -1,0 +1,14 @@
+package br.com.alfac.food.infra.item.persistence;
+
+import br.com.alfac.food.core.domain.item.CategoriaItem;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface ItemEntityRepository extends JpaRepository<ItemEntity, Long> {
+
+    List<ItemEntity> findByCategoria(CategoriaItem categoria);
+
+}
