+++ food-infra/src/main/java/br/com/alfac/food/infra/pagamento/persistence/PagamentoEntityRepository.java
@@ -1,0 +1,12 @@
+package br.com.alfac.food.infra.pagamento.persistence;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface PagamentoEntityRepository extends JpaRepository<PagamentoEntity, Long> {
+
+    Optional<PagamentoEntity> findByPedidoId(Long pedidoId);
+}
