+++ food-infra/src/main/resources/db/V1_0__setup_create_table_cliente.sql
