+++ food-infra/src/main/resources/db/V1_1__setup_create_table_item.sql
