+++ food-infra/src/main/resources/db/V1_2__alter_table_cliente_add_column_uuid.sql
