+++ food-main/pom.xml
@@ -27,11 +27,7 @@
             <artifactId>food-infra</artifactId>
             <version>0.0.1-SNAPSHOT</version>
         </dependency>
-        <dependency>
-            <groupId>br.com.alfac</groupId>
-            <artifactId>food-database</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-        </dependency>
+
         <dependency>
             <groupId>br.com.alfac</groupId>
             <artifactId>food-client</artifactId>
