+++ food-main/src/main/java/br/com/alfac/food/config/RepositoryConfiguration.java
@@ -1,9 +1,9 @@
 package br.com.alfac.food.config;
 
-import br.com.alfac.food.database.cliente.gateways.ClienteRepository;
-import br.com.alfac.food.database.cliente.mapper.ClienteEntityMapper;
-import br.com.alfac.food.database.cliente.persistence.ClienteEntityRepository;
 
+import br.com.alfac.food.infra.cliente.gateways.ClienteRepository;
+import br.com.alfac.food.infra.cliente.mapper.ClienteEntityMapper;
+import br.com.alfac.food.infra.cliente.persistence.ClienteEntityRepository;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
