+++ pom.xml
@@ -8,7 +8,6 @@
         <module>food-main</module>
         <module>food-client</module>
         <module>food-core</module>
-        <module>food-database</module>
         <module>food-infra</module>
     </modules>
     <parent>
